changes to beatmap manager and added some new scripts and files
extracting the drums from the MIDI file and putting it into a list works now. but every time you call it it does need a second, need to look into that further.

changed some variable names to be more logical.

tried to make a notes script but I need to aproach it differently. also made a path and songmanager script but songmanager probably isn't needed.

reach for the stars SO has the correct data now. MIDI files need to be in StreamingAssets (I Think)