Changed how beatmapmanager worked
beatmapmanager now distributes the notes to the desired drum kit. U need a reference to a gameObject with a path script for it to work. it also spawns notes now based on the time. Need to make it so the notes get on the drumkit the moment it needs to be hit instead of spawning the note the moment it needs to be hit

the path script holds a list of notes that it wil play. and you can assign note numbers in the inspector