changed how the timing works in beatmapmanager.
timing is now based on a "global timer" so it can start before the audiosource has begun playing. the reason for this is so the notes can have time to move towards the drums and stay in beat.

previously used the timer on the audiosource but that timer cant reach below 0 and that messed up some note timings in the script