changed how beatmapmanager works and prefabs
instead of creating a new instance that holds data, it now creates a note object itself and enables the object whenever it is time.  Need to add some note logic on when to count as a miss and need to double check the collision checks.

disabled every note prefab bc of the reason above.

added the drum class to every drum part