some big issues regarding note detection
Note detection apparently doesn't work and tried to fix it but still no luck. even tried chat gpt... but I think it got something to do with the index in the drums component and how the notes get removed from the list. gotta clean up a lot of code, will probably be behind in schedule.