I LOVE PROGRAMMING (This is canpaii btw we swapped laptops for a second )
the beatmap manager removed the notes from the list whenever it was time to spawn it in, but we wanted to acces those removed notes for the hit detection. So I simply removed that line of code from the beatmap manager and that worked out.
the note script now has a bool because if it doesnt get hit and gets destroyed because of a timer, it needs to count as a miss.
the score in scoremanager doesnt work yet it only counts by 1 from every perfect hit it gets. so I need to change the formula