Begin for pedal inputs
Made a new inout action with a pedal actionmap. There is one action now and the binding is any keyboard press because the pedal I use is identified as a keyboard with the letter "B" as input. I do want to make it so you can listen to any other pedal device and use that as a pedal and not just make it work for this one specific pedal. (that is also why I ordered two of them)