Made a start to the pedal input.
My pedal works now with the kick in my script test scene. I need to make it so it is easy to rebind for other pedals and/or inputs.

I did struggle a bit with the new input system, so it's definitely not perfect.

I also need to add more comments on my scripts. I just totally neglected that after the first week.