Beginning to drumroll script
the beatmapmanager can now detect streams of notes based on a threshhold, and sets the drum rollable bool to true on the note component.
I only need to make it so whenever you press a button on your vr controller it wil activate the drumroll and match the speed of the beat.