UI state machine and a little more
Made allot of progress with the UI. I didn't know how I wanted the backend to look like but I eventually made the choice to make it a statemachine and it took me a little longer to make most of it.

There are 4 main UI screens: Level Selector, Options, Endscreen and the main menu. Every screen (expect main menu for now) got their own script with their own functionality. The options isn't finished yet. that still  needs their logic adjusted and some other function aswell.

The StateMachine script sets the state of the game in one of 5 states. Main menu, level Select, Options, InGame and EndOfLeven, It changes what panels and drum portraits it needs to show.

I changed the drums component so it uses unity events based on what state the game is in. I felt like this was the appropriate way to do it.

I also added a music and SFX group in the audiomixxer but the adiosettings/setting in general aren't done yet.

there is still an issue where when you hit the snare at the beginning of the song it gives a few points for free but couldn't find the issue yet.

I added a few test texts and one button for testing in the main scene, they are pretty noticable so they wont  be missed easily. Also made a beautifull main menu panel