Audio Settings and Calibration and tried to add new song.
Audio settings and calibration works now. We just need a panel for the settings menu and make it clear what specific setting you are changing by adding a border or color shifts.

I tried to change the input handler so it is more dynamic. I want to make it in a way that you can add more inputs and events easier (that I will need for the drum roll later). But it doesnt work yet the way it is now, will need to look into it later.

I tried to add "world is mine" - ryo (supercell) into the game, but the MIDI file is really offsync and the MIDI file didn't work in FL studio for some reason, so I couldn't sync it with the song in there (the file itself works fine in other DAW). The song also doesnt stop, it plays on repeat.