Changed some note rotations and scrapped the pause
Adding in a pause now would probably take a little longer to add than I want it to, so I wanted to focus on other small things first such as fixxing our new UI so it shows the highscore.

Since you cant use a bas pedal while playing stand alone vr, we made the bass into the primary button of the left controller but will need to change that so also the right controller works properly.

also In one of my earlier pushes, I changed how the beatmapmanager locates the MIDI files. I honestly just used chat gpt for that since we only had a few days left for polishing and it was an issue that prevented us from playing the game in .apk format. It has something to do with the fact that .apk handles the streamingassets folder differently than .exe. It now uses unity web requests that I wasn't able to read into and understand yet.