The finishing touches (I hope)
I cant patch and run anymore for some reason so pushing this to try on a different laptop.

I made some very disqusting finishing touches, the new map selector script works a little weird so I had to make a new function so you can reset the highscore UI so it shows the apropriate highscore.

for performance reasons we were toying around with some player and urp settings, it felt like it helped a little bit but might just be placebo.

We deleted a few unecessary things in the scene, but not from all the scripts. I did that so the idea behind some gameMechanics such as the drum roll can still be seen (even though I never pushed most of the scripts I did for that mechanic since it wouldn't work how I wanted it to ;p)

we changed some audio clip settings to overide on android, because on the .apk most songs wouldn't work anymore (even though others did) and hopefully that would fix it.

I added a credits panel and changed the tekst positioning of some other UI panels. also fixxed a minor issue where some drum particles wouldn't properly play on the right transform