Final Push
This is the final push (unless I find a weird bug while trying to get some gameplay footage while playing the game).

I have Build 2 versions of this game, one .apk build with only 3 songs, since those are the 3 songs that had the most effort put into them with particles and the environment (and this build is for our teacher), and one build that is an .exe that we want to use for our up coming play expo. The .exe has a few more songs and is the build that we are pushing right now. Some songs have some syncing issue's or other small problems but we still wanted to add them so people still have the chance to play them.

The changes I made were mostly small bug fixes and UI related
:
Fixed an issue where the hit feedback text ("perfect", "early", etc..) wouldn't properly show after playing a few songs back to back.
fixed an issue where the crash cymbal would instantly play a song even though you were in the main menu or exit screen. (it had 2 mesh colliders)
Added an icon for the game.
fine tuned some UI text's.